[data] split test_all_to_all.py
