+++ python/ray/data/BUILD
@@ -297,9 +297,65 @@ py_test(
 )
 
 py_test(
-    name = "test_all_to_all",
+    name = "test_agg_e2e",
     size = "enormous",
-    srcs = ["tests/test_all_to_all.py"],
+    srcs = ["tests/test_agg_e2e.py"],
+    tags = [
+        "exclusive",
+        "team:data",
+    ],
+    deps = [
+        ":conftest",
+        "//:ray_lib",
+    ],
+)
+
+py_test(
+    name = "test_groupby_e2e",
+    size = "enormous",
+    srcs = ["tests/test_groupby_e2e.py"],
+    tags = [
+        "exclusive",
+        "team:data",
+    ],
+    deps = [
+        ":conftest",
+        "//:ray_lib",
+    ],
+)
+
+py_test(
+    name = "test_random_e2e",
+    size = "enormous",
+    srcs = ["tests/test_random_e2e.py"],
+    tags = [
+        "exclusive",
+        "team:data",
+    ],
+    deps = [
+        ":conftest",
+        "//:ray_lib",
+    ],
+)
+
+py_test(
+    name = "test_repartition_e2e",
+    size = "enormous",
+    srcs = ["tests/test_repartition_e2e.py"],
+    tags = [
+        "exclusive",
+        "team:data",
+    ],
+    deps = [
+        ":conftest",
+        "//:ray_lib",
+    ],
+)
+
+py_test(
+    name = "test_unique_e2e",
+    size = "enormous",
+    srcs = ["tests/test_unique_e2e.py"],
     tags = [
         "exclusive",
         "team:data",
